Cleanup code using a class-based abstraction
