+++ lib/index.ts
@@ -1,0 +1,3 @@
+export * from './IOstrichStoreNative';
+export * from './OstrichStore';
+export * from './utils';
